sql: fix confusing error message for create trigger in transaction
Previously when running create trigger in a transaction with `autocommit_before_ddl`
set to off, we would get an error message stating
`ERROR: CREATE TRIGGER is only implemented in the declarative schema changer`.
With these code changes we get a new with a hint that gives the suggestion of
trying to run in an implicit transaction and with a relevant link to the docs.

Fixes: #140053
Release note: Updated the `CREATE TRIGGER` `only implemented in the declarative schema changer`
 error message to include a helpful suggestion and link to relevant docs.
