+++ pkg/sql/BUILD.bazel
@@ -40,7 +40,6 @@ go_library(
         "comment_on_index.go",
         "comment_on_schema.go",
         "comment_on_table.go",
-        "comment_on_type.go",
         "compact_sql_stats.go",
         "completions.go",
         "conn_executor.go",
@@ -69,7 +68,6 @@ go_library(
         "create_stats.go",
         "create_table.go",
         "create_tenant.go",
-        "create_trigger.go",
         "create_type.go",
         "create_view.go",
         "created_sequence.go",
@@ -106,7 +104,6 @@ go_library(
         "drop_sequence.go",
         "drop_table.go",
         "drop_tenant.go",
-        "drop_trigger.go",
         "drop_type.go",
         "drop_view.go",
         "error_hints.go",
@@ -274,6 +271,7 @@ go_library(
         "txn_state.go",
         "type_change.go",
         "unary.go",
+        "unimplemented.go",
         "union.go",
         "unlisten.go",
         "unsplit.go",
