+++ pkg/sql/drop_owned_by.go
@@ -8,29 +8,13 @@ package sql
 import (
 	"context"
 
-	"github.com/cockroachdb/cockroach/pkg/server/telemetry"
 	"github.com/cockroachdb/cockroach/pkg/sql/sem/tree"
-	"github.com/cockroachdb/cockroach/pkg/sql/sqltelemetry"
-	"github.com/cockroachdb/cockroach/pkg/util/errorutil/unimplemented"
 )
 
 // dropOwnedByNode represents a DROP OWNED BY <role(s)> statement.
 type dropOwnedByNode struct {
 	// TODO(angelaw): Uncomment when implementing - commenting out due to linting error.
 	//n *tree.DropOwnedBy
-}
-
-func (p *planner) DropOwnedBy(ctx context.Context) (planNode, error) {
-	if err := checkSchemaChangeEnabled(
-		ctx,
-		p.ExecCfg(),
-		"DROP OWNED BY",
-	); err != nil {
-		return nil, err
-	}
-	telemetry.Inc(sqltelemetry.CreateDropOwnedByCounter())
-	// TODO(angelaw): Implementation.
-	return nil, unimplemented.NewWithIssue(55381, "drop owned by is not yet implemented")
 }
 
 func (n *dropOwnedByNode) startExec(params runParams) error {
